Refactor log cleanup and pre-config initialization
Moved log cleanup logic to a new cleanup_tasks service for better separation of concerns. Updated insert_page and pre_config to use the new cleanup function. Pre-config now also cleans up database logs on start. Minor improvements to login_page to initialize pre-config after login.
